requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine learning
lightgbm>=4.0.0
xgboost>=2.0.0

# API and web requests
requests>=2.31.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (for evaluation)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development and linting
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: Database support
# psycopg2-binary>=2.9.0
# SQLAlchemy>=2.0.0

# Optional: Advanced time series
# statsmodels>=0.14.0
# prophet>=1.1.0

# Optional: Model deployment
# fastapi>=0.100.0
# uvicorn>=0.23.0
